+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  - **Color Animator Parameter**: Manipulates quadruples of float parameter in Animators as Colors.
+ - **Vector4 Animator Parameter**: Manipulates quadruples of float parameters in Animaters as Vector4s.
 
 ## [0.2.0] - 2023-02-12